Added Trigger for the Yamaha XTZ750
This is a trigger I`ve been riding around with for about a year based on the Vmax one from @RempageR1 . The XTZ flywheel has 4 unevenly sized lobes, where one is the wide lobe used for triggering. This trigger therefore has a routine to note the times when an upward and downward tooth edge is seen and can therefore calculate the tooth width. The actual trigger needs to happen on the upward tooth edge so this is included in the routine.

IMPORTANT: The VR conditioner must have the zero-crossing disabled in order to have reliable triggering

This is my first commit to the Speeduino project, please let me know if there is anything I need to change or of you need any additional information. Thank you!